Upgrading to Teem 1.9. Main changes include API changes for var-arg functions and more careful use of unsigned types.  See http://sourceforge.net/project/shownotes.php?release_id=380683&group_id=85445 for more information.

+++ Core/Datatypes/NrrdData.cc
@@ -240,7 +240,7 @@ void NrrdData::io(Piostream& stream)
       stream.begin_cheap_delim();
       stream.io(nrrd->dim);
 		
-      int nrrddims[NRRD_DIM_MAX]; // defined in nrrd.h
+      size_t nrrddims[NRRD_DIM_MAX]; // defined in nrrd.h
       for (int p = 0; p<nrrd->dim; p++)
       {
 	stream.io(nrrddims[p]);
