+++ Packages/Teem/Dataflow/Modules/Unu/UnuReshape.cc
@@ -121,7 +121,7 @@ void
     return;
   }
 
-  int *sz = new int[nin->dim];
+  size_t *sz = new size_t[nin->dim];
   // Size/samples
   i=0, start=0;
   int which = 0, end=0, counter=0;
